todo-phase-III/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
pydantic[email]==2.5.0
email-validator>=2.0.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pytz==2023.3.post1
python-dotenv>=1.0.0
psutil>=5.9.0
psycopg2-binary>=2.9.9
